cctv/requirements.txt
openai==1.50.2
requests==2.32.5
python-dotenv==1.0.1
tenacity==9.1.2
